tests/submission/W3C/domhtml/anchor02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-a-charset" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<p>
<a charset="US-ASCII" href="../images/yellow.png">View Submit Button</a>
</p>


<script>
var domtest = async_test("The IDL attribute of the a element must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcharset;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcharset = testNode.charset;

      assert_equals(vcharset,"US-ASCII");
       
    domtest.done();
}));

</script>
<p>The charset IDL attribute of the a element must reflect the respective content attribute of the same name.</p>
<div id="log"></div>
</body></html>
